SAM23D/requirements.txt
# SAM23D Requirements
# Image Processing and Deep Learning
torch==2.1.0
torchvision==0.16.0
pillow>=10.0.0
numpy>=1.24.3

# 3D Point Cloud Processing
laspy==2.4.1
plyfile==1.0.0

# SAM2 Dependencies
hydra-core==1.3.2
omegaconf>=2.3.0
iopath==0.1.10

# Utilities
tqdm>=4.65.0
